pyproject.toml
[tool.poetry]
name = "thespycatagencytesttask"
version = "0.1.0"
description = ""
authors = ["Ivan Berestianyi <berestvania@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
